ir+metainterp: move PtrInfo variant structs to majit-ir::ptr_info
13 pure-data variant struct definitions land in `majit-ir::ptr_info`:
StrPtrInfo, VStringVariant + VStringPlainInfo / VStringSliceInfo /
VStringConcatInfo, VirtualInfo, VirtualArrayInfo, InstancePtrInfo,
StructPtrInfo, ArrayPtrInfo, VirtualStructInfo, VirtualArrayStructInfo,
VirtualRawSliceInfo, VirtualRawBufferInfo (+ its leaf-only impl block),
VirtualizableFieldState.

StrPtrInfo's three OptContext-coupled methods (getstrlen,
get_constant_string_spec, strgetitem) become `StrPtrInfoExt` extension
trait in metainterp, mirroring the IntBoundMakeGuards pattern. The
leaf `is_virtual()` method moves with the struct since it reads only
the embedded VStringVariant.

PtrInfo enum itself still lives in metainterp; this slice just makes
its variant payloads majit-ir-local in preparation.
